Huge refactoring to base code generation around newly introduced VerbPhrases.
VerbPhrases now can mediate defiinition of slots in the impl struct.

The previously implemented "open coding" form of inheritance is now
broken.  It might be supported by a new verb in the future.

Significant work was also done in the area of package imports and
naming.

New verbs: delete and discriminate.